backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langgraph==0.1.14
langchain==0.2.5
langchain-core>=0.2.22,<0.3
langchain-community==0.2.5
httpx==0.27.0
python-dotenv==1.0.1
supabase==2.4.6
redis==5.0.4
websockets==12.0
pydantic==2.7.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
pytest==8.2.2
pytest-asyncio==0.23.7
faker==25.3.0
beautifulsoup4==4.12.3
pytrends==4.9.2
requests==2.32.3
lxml==5.2.2
PyPDF2==3.0.1
pdfplumber>=0.10.0
lightgbm>=4.3.0
xgboost>=2.0.0
Pillow>=10.0.0
pytesseract==0.3.13
pdf2image>=1.16.0
rapidocr-onnxruntime==1.4.4
